Перенасла разметки в соответсвующие VIEW Добавила отсутпы где нужно
