+++ src/view/point-add-view.js
@@ -166,19 +166,21 @@ const createPointAddTemplate = () => `
 `;
 
 export default class PointAddView {
-  getTemplate() {
+  #element = null;
+
+  get template() {
     return createPointAddTemplate();
   }
 
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
 
-    return this.element;
+    return this.#element;
   }
 
   removeElement() {
-    this.element = null;
+    this.#element = null;
   }
 }
